requirements.txt
azure-functions
azure-storage-blob
azure-storage-queue
azure-storage-table
azure-identity